Update code,Fix the base type

+++ Carbon/src/Bootloader/boot.c
@@ -2,16 +2,16 @@
 #include "stddef.h"
 #include "stdio.h"
 typedef struct __multiboot_header__ {
-	uint64_t magic;
-	uint64_t flags;
-	uint64_t checksum;
+	uint32_t magic;
+	uint32_t flags;
+	uint32_t checksum;
 } MultibootHeader;
 MultibootHeader mbr __attribute__((section(".init.text"))) = {
 	.magic = 0x1BADB002,
 	.flags = 0x00000003,
 	.checksum = -(0x1BADB002 + 0x00000003),
 };
-uint64_t magic;
+uint32_t magic;
 Multiboot *multiboot_ptr;
 void boot_entry(void) {
 	asm volatile ("movl %%eax,%0" : "=m"(magic));
