+++ Carbon/src/Include/stddef.h
@@ -3,8 +3,8 @@
 #ifndef NULL
 #define NULL 0
 #endif // NULL
-typedef unsigned long uint64_t;
-typedef long int64_t;
+typedef unsigned long long uint64_t;
+typedef long long int64_t;
 typedef unsigned int uint32_t;
 typedef int int32_t;
 typedef unsigned short uint16_t;
